requirements.txt
requests
pandas
plotly_express
dash
dash-bootstrap-components
